asr/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
openai-whisper==20231117
webrtcvad==2.0.10
aiohttp==3.9.0
asyncio-mqtt==0.16.1
pydantic==2.5.0
python-multipart==0.0.6
mlflow==2.8.1
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
